refactor(common): make ProxyProtocol a top-level public class
Closes https://github.com/ooni/probe/issues/1905
